+++ R/download_predicate_dsl.R
@@ -201,7 +201,8 @@
 #' <https://www.gbif.org/developer/occurrence#predicates>
 #' @family downloads
 #' @examples
-#' pred("taxonKey", 5231190)
+#' # Uses COL (Catalogue of Life) Extended Release alpha-numeric keys by default
+#' pred("taxonKey", "Q2M4") # Calopteryx splendens
 #' pred_gt("elevation", 5000)
 #' pred_gte("elevation", 5000)
 #' pred_lt("elevation", 1000)
@@ -211,9 +212,9 @@
 #'   pred_gte("elevation", 5000))
 #' pred_or(pred_lte("year", 1989), pred("year", 2000))
 #' pred_and(pred_lte("year", 1989), pred("year", 2000))
-#' pred_in("taxonKey", c(2977832, 2977901, 2977966, 2977835))
+#' pred_in("taxonKey", c("Q2M4", "9WLSS", "Q2N2")) # COL XR alpha-numeric keys
 #' pred_in("basisOfRecord", c("MACHINE_OBSERVATION", "HUMAN_OBSERVATION"))
-#' pred_not(pred("taxonKey", 729))
+#' pred_not(pred("taxonKey", "Q2M4"))
 #' pred_like("catalogNumber", "PAPS5-560%")
 #' pred_notnull("issue")
 #' pred("basisOfRecord", "LITERATURE")
@@ -221,15 +222,15 @@
 #' pred("stateProvince", "California")
 #' pred("hasGeospatialIssue", FALSE)
 #' pred_within("POLYGON((-14 42, 9 38, -7 26, -14 42))")
-#' pred_or(pred("taxonKey", 2977832), pred("taxonKey", 2977901),
-#'   pred("taxonKey", 2977966))
-#' pred_in("taxonKey", c(2977832, 2977901, 2977966, 2977835))
+#' pred_or(pred("taxonKey", "Q2M4"), pred("taxonKey", "9WLSS"),
+#'   pred("taxonKey", "Q2N2"))
+#' pred_in("taxonKey", c("Q2M4", "9WLSS", "Q2N2", "Q2KZ"))
 #' pred("license", "CC_BY_4_0")
 #' pred_in("license", c("CC_BY_4_0", "CC_BY_NC_4_0"))
 
 #' @rdname download_predicate_dsl
 #' @export
-pred <- function(key, value, checklistKey = NULL) pred_factory("=")(key, value, checklistKey)
+pred <- function(key, value, checklistKey = "7ddf754f-d193-4cc9-b351-99906754a03b") pred_factory("=")(key, value, checklistKey)
 #' @rdname download_predicate_dsl
 #' @export
 pred_gt <- function(key, value) pred_factory(">")(key, value)
@@ -307,7 +308,7 @@ print.occ_predicate_list <- function(x, ...) {
 
 # helpers
 pred_factory <- function(type) {
-  function(key, value, checklistKey = NULL) {
+  function(key, value, checklistKey = "7ddf754f-d193-4cc9-b351-99906754a03b") {
     if (!length(key) == 1) stop("'key' must be length 1", call. = FALSE)
     if (!length(value) == 1) stop("'value' must be length 1", call. = FALSE)
     if (!is.null(checklistKey)) {
@@ -526,7 +527,7 @@ key_lkup <- list(
   INSTITUTION_KEY = "INSTITUTION_KEY"
   )
 
-parse_pred <- function(key, value, type = "and", checklistKey = NULL) {
+parse_pred <- function(key, value, type = "and", checklistKey = "7ddf754f-d193-4cc9-b351-99906754a03b") {
   assert(key, "character")
   assert(type, "character")
   assert(checklistKey, "character")
@@ -558,7 +559,7 @@ parse_pred <- function(key, value, type = "and", checklistKey = NULL) {
     list(type = unbox(type), parameter = unbox(key))
   } else if (type == "isNull") {
     list(type = unbox(type), parameter = unbox(key))
-  } else if (type == "equals" & !is.null(checklistKey) & key == "TAXON_KEY") {
+  } else if (type == "equals" && !is.null(checklistKey) && key == "TAXON_KEY") {
     list(
       type = unbox(type),
       key = unbox(key),
@@ -603,7 +604,7 @@ sub_str <- function(str, max = 100) {
   paste0(substring(str, 1, max), " ... ", sprintf("(N chars: %s)", nchar(str)))
 }
 parse_predicates <- function(user, email, type, format, verbatim_extensions, 
-  checklistKey = NULL, ...) {
+  checklistKey = "7ddf754f-d193-4cc9-b351-99906754a03b", ...) {
   tmp <- list(...)
   
   # Handle case where predicates are passed as positional args